PARQUET-2305: Allow Parquet to Proto conversion even though target schema has less fields (#1102)
If Parquet has any field which has been removed from the schema and Parquet to Proto conversion happens, it errors out due to Unknown fields. There could be some scenarios that we want to still convert PARQUET into the target proto schema object which has lesser fields. In this change using the ProtoParquetReader and specifying "ignoreUnknownFields" as an argument, the conversion can still be done which would ignore fields it can't convert and not error out.