PARQUET-2342: Fix writing corrupted parquet file by avoiding overflow on page value coun (#1135)
Parquet writer only checks the number of rows and the page size to
decide whether it needs to fit content written on a single page.
In the case of a composite column (ex: array/map) with many nulls,
it is possible to create 2billions+ values under the default page
size & row-count threshold (1MB, 20000 rows).
This change is adding an Integer limit threshold to avoid value count
overflow to happen within a single page.